requirements.txt
contourpy==1.0.7
cycler==0.11.0
filelock==3.10.2
fonttools==4.39.2
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.7.1
mizani==0.8.1
mpmath==1.3.0
networkx==3.0
numpy==1.24.2
packaging==23.0
palettable==3.3.0
pandas==1.5.3
patsy==0.5.3
Pillow==9.4.0
plotnine==0.10.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
scipy==1.10.1
six==1.16.0
statsmodels==0.13.5
sympy==1.11.1
torch==2.0.0
typing_extensions==4.5.0
tzdata==2022.7
